fix: address PR review feedback for avatar UI enhancements
- Add status line (Muted/Deafened/Online) to UserTooltip
- Rename 'Edit Avatar' to 'Edit Profile' in context menus
- Close avatar editor modal on disconnect
- Extract shared useTooltipPosition hook (DRY refactor)
- Import Tooltip.css in UserTooltip for portal styles
- Update plan/design docs per review comments
