+++ src/Brmble.Web/src/components/UserTooltip/UserTooltip.css
@@ -26,6 +26,12 @@
   text-overflow: ellipsis;
 }
 
+.user-tooltip-status {
+  font-family: var(--font-body);
+  font-size: var(--text-xs);
+  color: var(--text-muted);
+}
+
 .user-tooltip-comment {
   font-family: var(--font-body);
   font-size: var(--text-xs);
